pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "fungalflye"
version = "0.1.0"
description = "Long-read fungal genome assembly pipeline"
authors = [{name = "Jacob Durazo"}]
readme = "README.md"
requires-python = ">=3.9"

dependencies = [
    "typer",
    "pandas",
    "matplotlib",
    "biopython"
]

[project.scripts]
fungalflye = "fungalflye.wizard:wizard"